pets/requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.1.4
djangorestframework==3.14.0
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
